tools/comfyui/requirements.txt
dify_plugin~=0.0.1b52
websocket-client==1.7.0
